refactor: consolidate AI config/env/model core modules (Phase 1.5)
- Rename ai-config.ts → config.ts (config schema + read/write/clear only)
- Create ai-model.ts combining snapshot-api-config.ts + ai-setup-status.ts
- Move runAiConfigure and helpers into commands/ai.ts
- Delete snapshot-api-config.ts and ai-setup-status.ts
- Change source field from env:auto-openai → env:OPENAI_API_KEY style
- Update all imports and test assertions
