+++ packages/libretto/src/shared/run/browser.ts
@@ -11,7 +11,7 @@ import {
   SESSION_STATE_VERSION,
   SessionStateFileSchema,
 } from "../state/session-state.js";
-import { readLibrettoConfig } from "../../cli/core/ai-config.js";
+import { readLibrettoConfig } from "../../cli/core/config.js";
 
 async function pickFreePort(): Promise<number> {
   return await new Promise((resolve, reject) => {
